Move `instances` to the `core-common` package
