+++ core/common/src/main/scala/org/typelevel/otel4s/instances/LocalInstances.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.typelevel.otel4s.oteljava
+package org.typelevel.otel4s.instances
 
 import cats.Applicative
 import cats.Functor
@@ -23,19 +23,22 @@ import cats.effect.LiftIO
 import cats.effect.MonadCancelThrow
 import cats.mtl.Local
 
-object instances {
+trait LocalInstances {
+
   // We hope this instance is moved into Cats Effect.
-  implicit def localForIoLocal[F[_]: MonadCancelThrow: LiftIO, E](implicit
+  // See https://github.com/typelevel/cats-effect/pull/3429
+  implicit final def localForIOLocal[F[_]: MonadCancelThrow: LiftIO, E](implicit
       ioLocal: IOLocal[E]
   ): Local[F, E] =
     new Local[F, E] {
-      def applicative =
+      def applicative: Applicative[F] =
         Applicative[F]
-      def ask[E2 >: E] =
+      def ask[E2 >: E]: F[E2] =
         Functor[F].widen[E, E2](ioLocal.get.to[F])
       def local[A](fa: F[A])(f: E => E): F[A] =
         MonadCancelThrow[F].bracket(ioLocal.modify(e => (f(e), e)).to[F])(_ =>
           fa
         )(ioLocal.set(_).to[F])
     }
+
 }
