+++ core/common/src/test/scala/org/typelevel/otel4s/instances/LocalInstancesSuite.scala
@@ -1,0 +1,35 @@
+/*
+ * Copyright 2022 Typelevel
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.typelevel.otel4s.instances
+
+import cats.effect.IO
+import cats.effect.IOLocal
+import cats.effect.testkit.TestInstances
+import cats.mtl.laws.discipline.LocalTests
+import munit.DisciplineSuite
+import org.scalacheck.Arbitrary.arbString
+import org.typelevel.otel4s.instances.local._
+
+class LocalInstancesSuite extends DisciplineSuite with TestInstances {
+  implicit val ticker: Ticker = Ticker()
+
+  unsafeRun {
+    IOLocal("").map { implicit ioLocal =>
+      checkAll("IOLocal.LocalLaws", LocalTests[IO, String].local[String, Int])
+    }
+  }
+}
