+++ docs/instrumentation/tracing-java-interop.md
@@ -51,9 +51,9 @@ It can be constructed in the following way:
 import cats.effect._
 import cats.mtl.Local
 import cats.syntax.functor._
+import org.typelevel.otel4s.instances.local._ // brings Local derived from IOLocal
 import org.typelevel.otel4s.oteljava.context.Context
 import org.typelevel.otel4s.oteljava.OtelJava
-import org.typelevel.otel4s.oteljava.instances._ // brings Local derived from IOLocal
 import io.opentelemetry.api.GlobalOpenTelemetry
 
 def createOtel4s[F[_]: Async](implicit L: Local[F, Context]): F[OtelJava[F]] =
