+++ examples/src/main/scala/PekkoHttpExample.scala
@@ -37,9 +37,9 @@ import org.apache.pekko.http.scaladsl.server.Directives._
 import org.apache.pekko.http.scaladsl.server.Route
 import org.apache.pekko.util.ByteString
 import org.typelevel.otel4s.Attribute
+import org.typelevel.otel4s.instances.local._
 import org.typelevel.otel4s.oteljava.OtelJava
 import org.typelevel.otel4s.oteljava.context.Context
-import org.typelevel.otel4s.oteljava.instances._
 import org.typelevel.otel4s.trace.Tracer
 
 import scala.concurrent.Future
@@ -78,7 +78,7 @@ object PekkoHttpExample extends IOApp.Simple {
 
   def run: IO[Unit] =
     IOLocal(Context.root).flatMap { implicit ioLocal: IOLocal[Context] =>
-      implicit val local: Local[IO, Context] = localForIoLocal
+      implicit val local: Local[IO, Context] = localForIOLocal
       val otelJava: OtelJava[IO] = OtelJava.local(GlobalOpenTelemetry.get())
 
       otelJava.tracerProvider.get("com.example").flatMap {
