+++ oteljava/trace/src/test/scala/org/typelevel/otel4s/oteljava/trace/TracerSuite.scala
@@ -52,9 +52,9 @@ import org.typelevel.otel4s.context.propagation.{
   PassThroughPropagator => SPassThroughPropagator
 }
 import org.typelevel.otel4s.context.propagation.ContextPropagators
+import org.typelevel.otel4s.instances.local._
 import org.typelevel.otel4s.oteljava.context.Context
 import org.typelevel.otel4s.oteljava.context.propagation.PropagatorConverters._
-import org.typelevel.otel4s.oteljava.instances._
 import org.typelevel.otel4s.trace.Span
 import org.typelevel.otel4s.trace.Tracer
 import org.typelevel.otel4s.trace.TracerProvider
